+++ src/test/java/org/zalando/nakadi/controller/PartitionsControllerTest.java
@@ -16,7 +16,7 @@ import org.zalando.nakadi.domain.Storage;
 import org.zalando.nakadi.domain.Timeline;
 import org.zalando.nakadi.exceptions.InternalNakadiException;
 import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
-import org.zalando.nakadi.exceptions.ServiceUnavailableException;
+import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.repository.EventTypeRepository;
 import org.zalando.nakadi.repository.TopicRepository;
 import org.zalando.nakadi.repository.db.EventTypeCache;
@@ -142,7 +142,7 @@ public class PartitionsControllerTest {
     @Test
     public void whenListPartitionsAndNakadiExceptionThenServiceUnavaiable() throws Exception {
         when(timelineService.getActiveTimelinesOrdered(eq(TEST_EVENT_TYPE)))
-                .thenThrow(ServiceUnavailableException.class);
+                .thenThrow(ServiceTemporarilyUnavailableException.class);
 
         final ThrowableProblem expectedProblem = Problem.valueOf(SERVICE_UNAVAILABLE, null);
         mockMvc.perform(
@@ -260,7 +260,7 @@ public class PartitionsControllerTest {
     @Test
     public void whenGetPartitionAndNakadiExceptionThenServiceUnavaiable() throws Exception {
         when(timelineService.getActiveTimelinesOrdered(eq(TEST_EVENT_TYPE)))
-                .thenThrow(ServiceUnavailableException.class);
+                .thenThrow(ServiceTemporarilyUnavailableException.class);
 
         final ThrowableProblem expectedProblem = Problem.valueOf(SERVICE_UNAVAILABLE, null);
         mockMvc.perform(